backend/requirements-dev.txt
Django==1.11.24
djangorestframework==3.9.4
drf-nested-routers==0.91.0
Pillow==6.1.0
django-cors-middleware==1.4.0
django-webpack-loader==0.6.0